Refactored code internally to make it possible to provide a helper function that generates a density map without any input data
